+++ src/server/game/Globals/ObjectMgr.h
@@ -612,7 +612,7 @@ struct MailLevelReward
     MailLevelReward() : raceMask({ 0 }), mailTemplateId(0), senderEntry(0) { }
     MailLevelReward(uint32 _raceMask, uint32 _mailTemplateId, uint32 _senderEntry) : raceMask({ _raceMask }), mailTemplateId(_mailTemplateId), senderEntry(_senderEntry) { }
 
-    Trinity::RaceMask<uint32> raceMask;
+    Trinity::RaceMask<uint64> raceMask;
     uint32 mailTemplateId;
     uint32 senderEntry;
 };
